Replaced $show and construct()
Replacements for `construct` allow for difference, intersection, and hull operations to be taken over trees formed from nested primitives
`$show` is being replaced by $class, which will allow the aforementioned operations to be performed for primitives of a given class. This is intended to be used much like the "`$()`" selector function in jquery.